+++ example/__init__.py
@@ -13,15 +13,14 @@ def includeme(config):
         tkt_policy,
         Context1,
     )
-    config.register_authentication_policy(
-        tkt_policy,
-        Context2,
-    )
+
     config.register_authentication_policy(
         BasicAuthAuthenticationPolicy('realm'),
         Context2,
     )
+
     config.scan()
+
     config.commit()
 
 
@@ -45,6 +44,10 @@ class Context5(Context4):
     pass
 
 
-@authentication_policy(Context3, Context4)
+class ToBeRemovedPolicy(object):
+    pass
+
+
+@authentication_policy(contexts=(Context3, Context4))
 class Context345Policy(object):
     pass